backend/requirements.txt
# IP Agent Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Utils
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Database (using Supabase SDK instead)
# supabase 需要额外安装
